feat-前端商品搜索展示相关

+++ src/main/java/com/bxk/campusbazaar/api/service/Impl/UserServiceImpl.java
@@ -115,8 +115,8 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(int id) {
-        return userMapper.selectByPrimaryKey((long) id);
+    public User getUserById(long id) {
+        return userMapper.selectByPrimaryKey(id);
     }
 
     @Override
